Implement tap focus
Give the Canvas a name and add a Tap handler, lock focus parameters on
successful tap focus and release them when picture has been taken. Note
that taps and the region have to be converted into the coordinate space of
the preview image which is always 640x480 independent of the app
orientation. We use that as a clipping rectangle for the focus area to
avoid getting invalid argument errors due to the FocusRegion falling
outside the limits.