Fix for green viewfinder
A known issue affecting the camera apps on Lumia WP8 devices. To
workaround it (or at least limit the incidence) applications should
release the camera object as soon as they are no longer using it and
then create it again when needed.

In this case, the camera object is released before going to the
PreviewPage and re-created on return to the MainPage.